// src/app/main/notifications-list/shared/models/notifications-list.model.ts
export class NotificationsHistory {
  body: string;
  createdOn: string;
  type: string;
  proposedTime: Date;
  proposedTimeExact?: string;
  eventId: string;
  response: string;
  attendeeId: string;
  status: string;
  isOrganiser: boolean;
  notificationId: string;
  isRespondedTo: string;
  constructor() {}
}